requirements.txt
Django>=5.0
psycopg2-binary
djangorestframework
python-decouple
requests